+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # BlogMore ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added a table of longest posting streaks to the stats page, showing the
+  top 10 consecutive-day posting streaks of two or more days.
+  ([#369](https://github.com/davep/blogmore/pull/369))
+
 ## v2.10.0
 
 **Released: 2026-07-07**